chore: Refactor images API to with OpenAI compatible behaviour
